Handle implicit and accumulator address mode (devel).
These two modes are unique, because neither of them interact with RAM;
only CPU state will be modified. To manage implicit mode, we simply
consider all instructions to work in this mode; later when registering
an instruction we simply disable this mode. However for accumulator mode,
we need to somehow signal the instruction that it's working in such mode.
To do so, we simply use process status register's flag.